+++ R/point.R
@@ -1,64 +1,81 @@
-#' @title Simple point chart
+#' @title Simple bar chart
 #'
 #' @param df A data frame.
 #' @param x A numeric column.
 #' @param y A character column or coercible as a character column.
 #' @param group Some grouping categorical column, e.g. administrative areas or population groups.
-#' @param palette Palette name from 'pal_reach()'.
-#' @param reverse Boolean indicating whether the palette should be reversed.
-#' @param family The font family for all plot's texts. Default to "Leelawadee".
-#' @param alpha Transparency.
-#' @param size Dot size. Default to 1.5.
+#' @param flip TRUE or FALSE. Default to TRUE or horizontal bar plot.
+#' @param alpha Fill transparency.
+#' @param size Point size.
 #' @param x_title The x scale title. Default to NULL.
 #' @param y_title The y scale title. Default to NULL.
 #' @param group_title The group legend title. Default to NULL.
 #' @param title Plot title. Default to NULL.
 #' @param subtitle Plot subtitle. Default to NULL.
-#' @param caption Caption title string. Default to NULL.
-#' @param text_size The size of all text other than the title, subtitle and caption. Defaults to 10.
-#' @param title_size The size of the title text. Defaults to 14.
-#' @param legend_position Position of the legend; Default to "right". Can take "right", "left", "top", "bottom" or "none".
-#' @param ... Other arguments to be passed to "ggblanket::gg_col"
-#'
-#' @description `ggblanket` as internals for deciding whether the bar chart is horizontally readable.
+#' @param caption Plot caption. Default to NULL.
+#' @param theme Whatever theme. Default to theme_reach().
 #'
 #' @return A bar chart
 #'
 #' @export
-point_reach <- function(df, x, y, group = NULL, palette = "main", reverse = FALSE, family = "Leelawadee", alpha = 1, size = 1.5, x_title = NULL, y_title = NULL, group_title = NULL, title = NULL, subtitle = NULL, caption = NULL, text_size = 10, title_size = 14, legend_position = "right", ...){
+point <- function(df, x, y, group = NULL, flip = TRUE, alpha = 1, size = 1, x_title = NULL, y_title = NULL, group_title = NULL, title = NULL, subtitle = NULL, caption = NULL, theme = theme_reach()){
 
-  pal <- pal_reach(palette)
+  # To do :
+  # - automate bar width and text size, or at least give the flexibility and still center text
+  # - add facet possibility
 
-  if(is.null(pal)) rlang::warn(
-    c(paste0("There is no palette '", palette, "' for initiative 'reach'. Fallback to ggblanket's default color palette."),
-      "i" = paste0("Use `pal_reach(show_palettes = T)` to see the list of availabale palettes.")
+  # Prepare group, x and y names
+  # if (is.null(x_title)) x_title <- rlang::as_name(rlang::enquo(x))
+  # if (is.null(y_title)) y_title <- rlang::as_name(rlang::enquo(y))
+  # if (is.null(group_title)) group_title <- rlang::as_name(rlang::enquo(group))
+
+  # Mapping
+  g <- ggplot2::ggplot(
+    df,
+    mapping = ggplot2::aes(x = {{ x }}, y = {{ y }}, fill = {{ group }}, color = {{ group }}
     )
   )
 
-  pl <- df |>
-    ggblanket::gg_point(x = {{ x }},
-                      y = {{ y }},
-                      col = {{ group }},
-                      x_title = x_title,
-                      y_title = y_title,
-                      col_title = group_title,
-                      alpha = alpha,
-                      size = size,
-                      title = title,
-                      subtitle = subtitle,
-                      caption = caption,
-                      col_legend_place = legend_position,
-                      theme = theme_reach(
-                        palette = palette,
-                        reverse = reverse,
-                        family = family,
-                        text_size = text_size,
-                        title_size = title_size,
-                        plot_background_pal = "#FFFFFF",
-                        panel_background_pal = "#FFFFFF"
-                      ),
-                      ...
+  # Add title, subtitle, caption, x_title, y_title
+  g <- g + ggplot2::labs(
+    title = title,
+    subtitle = subtitle,
+    caption = caption,
+    x = x_title,
+    y = y_title,
+    color = group_title,
+    fill = group_title
+  )
+
+  width <- 0.5
+  dodge_width <- 0.5
+
+  # Should the graph use position_fill?
+  g <- g + ggplot2::geom_point(
+      alpha = alpha,
+      size = size
     )
 
-  return(pl)
+  # Labels to percent and expand scale
+  # if (percent) {
+  #   g <- g + ggplot2::scale_y_continuous(
+  #     labels         = scales::label_percent(
+  #       accuracy     = 1,
+  #       decimal.mark = ",",
+  #       suffix       = " %"),
+  #     expand = c(0.01, 0.1)
+  #   )
+  # } else {
+  #   g <- g + ggplot2::scale_y_continuous(expand = c(0.01, 0.1))
+  # }
+
+  # # Because a text legend should always be horizontal, especially for an horizontal bar graph
+  if (flip){
+    g <- g + ggplot2::coord_flip()
+  }
+
+  # Add theme
+  g <- g + theme
+
+  return(g)
 }